This was supposed to be ignored by .gitignore
This was supposed to be ignored by .gitignore, but I am cursed by Git and cause it to do completely opposite things to what it's supposed to do. So, consequently, the one directory that I told Git to ignore was the only directory Git decided to push to the repo. So, rather than deal with Git, I am deleting it through its nicer cousin Github, who actually gets along with me on Fridays and doesn't actively try to sabotage my work.